test(nuxt): split packed entry test into its own file
